requirements.txt
Django==1.10.0
djangorestframework==3.3.3
unidecode==0.4.19
psycopg2==2.6.1
django-filter==0.13.0
django-modeltranslation==0.11
bitstring==3.1.5